BigQuery permissions error is resolved. All injectors are working properly. Credential supplier from connector is integrated. Self writing test has been written at SparkBigQueryWriteTest.java and is passing all green!

+++ connector/src/test/java/com/google/cloud/spark/bigquery/v2/SparkBigQueryWriteTest.java
@@ -1,7 +1,9 @@
 package com.google.cloud.spark.bigquery.v2;
 
 import com.google.cloud.bigquery.BigQuery;
+import com.google.cloud.bigquery.BigQueryOptions;
 import com.google.cloud.bigquery.DatasetInfo;
+import com.google.cloud.bigquery.connector.common.BigQueryClient;
 import com.google.cloud.bigquery.testing.RemoteBigQueryHelper;
 import org.apache.log4j.Level;
 import org.apache.log4j.LogManager;
@@ -14,6 +16,11 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import java.util.Arrays;
+
+import static com.google.common.truth.Truth.assertThat;
+import static org.junit.Assert.fail;
+
 public class SparkBigQueryWriteTest {
 
     public static final Logger logger = LogManager.getLogger("com.google.cloud.spark.bigquery");
@@ -22,6 +29,13 @@ public class SparkBigQueryWriteTest {
     public static final String DATASET = "ymed_"+RemoteBigQueryHelper.generateDatasetName();
     public static final String TABLE = "testTable";
     public static final String DESCRIPTION = "Spark BigQuery connector write session tests.";
+
+    public static final String TEST_DATA_FRAME = /*"google.com:hadoop-cloud-dev:ymed.Titanic"*/
+            /*"bigquery-public-data:bbc_news.fulltext"*/
+            /*"bigquery-public-data:fda_drug.drug_enforcement"*/
+            /*"bigquery-public-data:san_francisco_bikeshare.bikeshare_station_info"*/
+            /*"bigquery-public-data:austin_incidents.incidents_2016"*/
+            "google.com:hadoop-cloud-dev:ymed.all_types";
 
     public static SparkSession spark;
     public static BigQuery bigquery;
@@ -37,15 +51,9 @@ public class SparkBigQueryWriteTest {
                 .master("local[*]")
                 .getOrCreate();
         df = spark.read().format("com.google.cloud.spark.bigquery.v2.BigQueryDataSourceV2")
-                .option("table", "google.com:hadoop-cloud-dev:ymed.Titanic"
-                        /*"bigquery-public-data:bbc_news.fulltext"*/
-                        /*"bigquery-public-data:fda_drug.drug_enforcement"*/
-                        /*"bigquery-public-data:san_francisco_bikeshare.bikeshare_station_info"*/
-                        /*"bigquery-public-data:austin_incidents.incidents_2016"*/
-                        /*"google.com:hadoop-cloud-dev:spark_bigquery_it_1588126384826.all_types"*/)
-                .load();
-        RemoteBigQueryHelper bigqueryHelper = RemoteBigQueryHelper.create();
-        bigquery = bigqueryHelper.getOptions().getService();
+                .option("table", TEST_DATA_FRAME)
+                .load().drop("ts").drop("dt").toDF(); // TODO: when timestamps are supported externally in Vortex, delete ".drop" operations.
+        bigquery = BigQueryOptions.getDefaultInstance().getService();
         DatasetInfo datasetInfo =
                 DatasetInfo.newBuilder(/* datasetId = */ DATASET).setDescription(DESCRIPTION).build();
         bigquery.create(datasetInfo);
@@ -54,10 +62,10 @@ public class SparkBigQueryWriteTest {
 
     @AfterClass
     public static void close() {
-        /*if (bigquery != null) {
+        if (bigquery != null) {
             RemoteBigQueryHelper.forceDelete(bigquery, DATASET);
             logger.info("Deleted test dataset: " + DATASET);
-        }*/
+        }
     }
 
     @Test
@@ -68,5 +76,17 @@ public class SparkBigQueryWriteTest {
                 .option("project", PROJECT)
                 .mode(SaveMode.Overwrite)
                 .save();
+
+        Dataset<Row> actual = spark.read()
+                .format("com.google.cloud.spark.bigquery.v2.BigQueryDataSourceV2")
+                .option("table", TABLE)
+                .option("dataset", DATASET)
+                .option("project", PROJECT)
+                .load();
+
+        assertThat(actual.schema()).isEqualTo(df.schema());
+
+        long intersectionRowCount = actual.intersectAll(df).count();
+        assertThat(intersectionRowCount == actual.count() && intersectionRowCount == df.count());
     }
 }
